Added Binary Search in dart language (In the dart folder)
Binary search is an efficient algorithm for finding an item from a sorted list of items. It works by repeatedly dividing in half the portion of the list that could contain the item, until you've narrowed down the possible locations to just one. Its worst case time complexity is O(log N).